requirements.txt
numpy==2.1.1
pandas==2.2.2
scikit-learn==1.5.1
scipy==1.14.1
fastprogress==1.0.3
torchsampler==0.1.2
matplotlib==3.9.2
seaborn==0.13.2
captum==0.7.0
statsmodels==0.14.3
bayesian-optimization==1.5.1
einops==0.8.0
prettytable==3.11.0
biopython==1.84
transformers==4.44.2